+++ components/shared/index.ts
@@ -1,6 +1,6 @@
 // Shared Components
 export { default as Header } from 'components/shared/Header';
 export { default as Footer } from 'components/shared/Footer';
-
+export { default as LargeBtn } from 'components/shared/Button/LargeBtn';
 // Only Styled Components
 export * as StyledLayout from 'components/shared/styled/layout';
